[QUAR-51] 📝 Rename: types 폴더명과 중복되는 파일명의 Types 제거
