+++ src/common/types/mailing.ts
