+++ src/components/types/checkFavoriteField.ts
