+++ src/pages/MyPage/View/AccountManageView.tsx
@@ -4,7 +4,7 @@ import ArrowDown from '@icon/icon-arrow-down.svg?react';
 import ToggleON from '@icon/icon-toggle-on.svg?react';
 import ToggleOFF from '@icon/icon-toggle-off.svg?react';
 import useUserService from '@feature/User/useUserService.ts';
-import { IUserDataProps } from '@type/userTypes.ts';
+import { IUserDataProps } from '@type/user.ts';
 import '../style/accountManageView.scss';
 
 const AccountManageView = ({ data }: IUserDataProps) => {
